Added load_daq and load_exper functions.  Refactored load_file to use the new functions.
I had a problem loading the experiment protocol data.  Different
experiments have different protocols with different formats.  So, I had
to find a more general method for loading in the data.  The data is
some struct with a bunch of variable fields.  It records the fields,
and if one is 'param', looks through the fields there too.  It gets the
data types of the fields too, and uses the field names and data types
to build a record array.